+++ src/transform_component.rs
@@ -1,9 +1,9 @@
 use bevy_ecs::component::Component;
 use nalgebra::{Quaternion, Vector3};
 
-#[derive(Default, Component)]
+#[derive(Default, Component, Debug)]
 pub struct TransformComponent {
-    pub position: Vector3<f64>,
-    pub rotation: Quaternion<f64>,
-    pub scale: Vector3<f64>,
+    pub position: Vector3<f32>,
+    pub rotation: Quaternion<f32>,
+    pub scale: Vector3<f32>,
 }
