+++ src/velocity_component.rs
@@ -1,8 +1,8 @@
 use bevy_ecs::component::Component;
-use nalgebra::{Quaternion, Vector3};
+use nalgebra::Vector3;
 
 #[derive(Default, Component)]
 pub struct VelocityComponent {
-    pub translational: Vector3<f64>,
-    pub angular: Vector3<f64>,
+    pub translational: Vector3<f32>,
+    pub angular: Vector3<f32>,
 }
